test: Remove node-mocks-http references from integration tests
