--- test/assets/permanent.txt
@@ -1,1 +1,0 @@
-TEST
