--- test/assets/testfile0.txt
@@ -1,1 +1,0 @@
-TESTFILE0
