--- test/assets/testfile1.txt
@@ -1,1 +1,0 @@
-TESTFILE1
