--- test/assets/testfile2.txt
@@ -1,1 +1,0 @@
-TESTFILE2
